Commit changed how temperature is displayed

+++ WeatherApp/models/TemperaturesMaxModel.swift
@@ -10,9 +10,9 @@ import SwiftyJSON
 
 class TemperaturesMaxModel{
    
-    let temperature: Double
+    let temperature: Int
     
     init(json:JSON){
-        self.temperature = json[].doubleValue
+        self.temperature = json[].intValue
     }
 }
